perf: async confirmation + parallel chunks for transfer batches
Split the 1.1k-line transfer-batches handler into focused modules
(resolve/transaction/estimate/execute/create/respond/read/policy).

- Submit chunks concurrently and respond without awaiting on-chain
  confirmation; the pending-transfers job settles recipient rows and
  the parent batch (settle-first so failures self-heal on retry)
- Fetch wallet policy rows once per batch instead of per recipient
- Replace the per-transfer response fan-out with listTransfersByIds
- Share batch status derivation via deriveTransferBatchStatus
- Sum amounts with @sdp/payments/decimal fixed-point helpers
- Add cron settlement, 500-recipient scaling, and concurrent-burst tests
